move update fn from store to query folder
