+++ editor.planx.uk/src/pages/FlowEditor/components/Team/queries/createAndAddUserToTeam.tsx
@@ -2,6 +2,7 @@ import { FetchResult, gql } from "@apollo/client";
 import { GET_USERS_FOR_TEAM_QUERY } from "routes/teamMembers";
 
 import { client } from "../../../../../lib/graphql";
+import { AddNewEditorFormValues } from "../types";
 
 type CreateAndAddUserResponse = FetchResult<{
   insert_users_one: { id: number; __typename: "users" };
@@ -45,8 +46,38 @@ export const createAndAddUserToTeam = async (
       { query: GET_USERS_FOR_TEAM_QUERY, variables: { teamSlug } },
     ],
   });
+
   if (response.data) {
     return response.data.insert_users_one;
   }
   throw new Error("Unable to create user");
+};
+
+export const updateTeamMember = async (
+  userId: number,
+  userValues: AddNewEditorFormValues,
+) => {
+  const response = await client.mutate({
+    mutation: gql`
+      mutation UpdateUser($userId: Int, $userValues: users_set_input) {
+        update_users(where: { id: { _eq: $userId } }, _set: $userValues) {
+          returning {
+            id
+          }
+        }
+      }
+    `,
+    variables: {
+      userId: userId,
+      userValues: {
+        first_name: userValues.firstName,
+        last_name: userValues.lastName,
+        email: userValues.email,
+      },
+    },
+  });
+  if (response.data) {
+    return response.data.update_users;
+  }
+  throw new Error("Unable to update user");
 };
